--- src/DD_tools/mam_ansp_fix/README.md
@@ -1,5 +1,0 @@
-tool to fix the duplication issue with `mam.ansp.org` server from gbif
-
-Additional config fields:
-
-* `uuid_table_path` - path to uuid table with duplicated uuids
